@@ -2,7 +2,6 @@ from typing import Dict, Optional
 
 import jax_dataloader as jdl
 from cryojax.data import (
-    AbstractParticleStackDataset,
     RelionParticleStackDataset,
 )
 from jaxtyping import PRNGKeyArray
@@ -22,7 +21,7 @@ class CustomJaxDataset(jdl.Dataset):
         self.cryojax_dataset = cryojax_dataset
         self.per_particle_args = per_particle_args
 
-    def __getitem__(self, index) -> Dict[str, AbstractParticleStackDataset | PerParticleArgs]:
+    def __getitem__(self, index) -> Dict[str, Dict | PerParticleArgs]:
         if self.per_particle_args is None:
             per_particle_args = None
         else:
