@@ -2,6 +2,6 @@ from ._datasets_and_dataloaders import (
     create_dataloader as create_dataloader,
 )
 from ._relion_data_generator import (
-    generate_relion_parameter_dataset as generate_relion_parameter_dataset,
+    generate_relion_parameter_file as generate_relion_parameter_file,
     simulate_relion_dataset as simulate_relion_dataset,
 )
