@@ -1,17 +1,14 @@
-from typing import Tuple
+from typing import Dict, Tuple
 
 import cryojax.simulator as cxs
 import jax.numpy as jnp
-from cryojax.data import (
-    RelionParticleParameterFile,
-)
 from cryojax.image.operators import AbstractMask
 from cryojax.inference.distributions import IndependentGaussianPixels
 from jaxtyping import Array, Float, Int, PRNGKeyArray
 
 
 def render_image_with_white_gaussian_noise(
-    relion_particle_parameters: RelionParticleParameterFile,
+    particle_parameters: Dict,
     constant_args: Tuple[
         Tuple[cxs.AbstractPotentialRepresentation],
         cxs.AbstractPotentialIntegrator,
@@ -27,7 +24,7 @@ def render_image_with_white_gaussian_noise(
     and noise variance. The noise is White Gaussian noise.
 
     **Arguments:**
-        relion_particle_parameters: The particle parameters.
+        particle_parameters: The particle parameters.
         constant_args: A tuple with the potential and potential integrator.
         per_particle_args: A containing a random jax key,
             the potential_idx to use, and the noise variance.
@@ -40,18 +37,18 @@ def render_image_with_white_gaussian_noise(
 
     structural_ensemble = cxs.DiscreteStructuralEnsemble(
         potentials,
-        relion_particle_parameters.pose,
+        particle_parameters["pose"],
         cxs.DiscreteConformationalVariable(potential_idx),
     )
 
     scattering_theory = cxs.WeakPhaseScatteringTheory(
         structural_ensemble,
         potential_integrator,
-        relion_particle_parameters.transfer_theory,
+        particle_parameters["transfer_theory"],
     )
 
     image_model = cxs.ContrastImageModel(
-        relion_particle_parameters.instrument_config, scattering_theory, mask=mask
+        particle_parameters["instrument_config"], scattering_theory, mask=mask
     )
 
     distribution = IndependentGaussianPixels(
